requirements.txt
# Document Processing with Docling
docling>=2.61.0
pillow>=10.2.0

# OCR fallback
pytesseract>=0.3.10

# AI Integration
ollama>=0.1.6
requests>=2.32.2

# Markdown Processing
markdown>=3.5.2
python-frontmatter>=1.1.0

# Utilities
python-magic>=0.4.27
pyyaml>=6.0.1
tqdm>=4.66.1
colorama>=0.4.6

# Logging
loguru>=0.7.2

# PowerPoint image extraction fallback
python-pptx>=0.6.21

# CSV/Spreadsheet support
pandas>=2.0.0
tabulate>=0.9.0
openpyxl>=3.1.0

# Email support
extract-msg>=0.45.0

# Web API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0

